refactored RadioStreamPreference, created dialog class reusable outside of the preference
