+++ src/com/firebirdberlin/nightdream/ui/RadioStreamDialogListener.java
@@ -1,0 +1,7 @@
+package com.firebirdberlin.nightdream.ui;
+
+import com.firebirdberlin.radiostreamapi.models.RadioStation;
+
+public interface RadioStreamDialogListener {
+    void onRadioStreamSelected(RadioStation station);
+}
